+++ ampltools/ampltools/modules/__init__.py
@@ -1,11 +1,14 @@
 # -*- coding: utf-8 -*-
-__version__ = "0.4.0"
+__version__ = "0.4.7b0"
 
 from .amplpypi import (
     path,
+    run,
+    activate,
     install_modules as install,
     uninstall_modules as uninstall,
     load_modules as load,
+    preload_modules as preload,
     installed_modules as installed,
     available_modules as available,
     generate_requirements as requirements,